generate sample mcts kids
