+++ checkstyle.xml
@@ -1,6 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE module PUBLIC "-//Puppy Crawl//DTD Check Configuration 1.2//EN" "http://www.puppycrawl.com/dtds/configuration_1_2.dtd">
 <module name="Checker">
+    <module name="SuppressionFilter">
+        <property name="file" value="suppressions.xml"/>
+    </module>
     <module name="SuppressionCommentFilter"/>
     <module name="SuppressWithNearbyCommentFilter">
         <property name="commentFormat" value="NO CHECKSTYLE (\w+)"/>
