+++ suppressions.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0"?>
+
+<!DOCTYPE suppressions PUBLIC "-//Puppy Crawl//DTD Suppressions 1.1//EN" "http://www.puppycrawl.com/dtds/suppressions_1_1.dtd">
+
+<suppressions>
+    <suppress checks="MagicNumber" files="[/\\]testing[/\\]src[/\\]"/>
+</suppressions>
