+++ testing/src/main/java/org/motechproject/nms/testing/service/TestingService.java
@@ -12,4 +12,6 @@ public interface TestingService {
     void createSubscriptionPacks();
 
     String createMctsMoms(int count) throws IOException;
+
+    String createMctsKids(int count) throws IOException;
 }
